+++ chapter08/02_vulkan_collisions/vulkan/VkRenderer.cpp
@@ -3713,7 +3713,7 @@ void VkRenderer::reactToInstanceCollisions() {
 
 void VkRenderer::runComputeShaders(std::shared_ptr<AssimpModel> model, int numInstances,
     uint32_t modelOffset, uint32_t instanceOffset, bool useEmptyBoneOffsets) {
-  uint32_t numberOfBones = model->getBoneList().size();
+  uint32_t numberOfBones = static_cast<uint32_t>(model->getBoneList().size());
 
   /* node transformation */
   vkCmdBindPipeline(mRenderData.rdComputeCommandBuffer, VK_PIPELINE_BIND_POINT_COMPUTE,
@@ -3782,7 +3782,7 @@ void VkRenderer::runComputeShaders(std::shared_ptr<AssimpModel> model, int numIn
 
 void VkRenderer::runBoundingSphereComputeShaders(std::shared_ptr<AssimpModel> model, int numInstances,
     uint32_t modelOffset) {
-  uint32_t numberOfBones = model->getBoneList().size();
+  uint32_t numberOfBones = static_cast<uint32_t>(model->getBoneList().size());
 
   /* node transformation */
   vkCmdBindPipeline(mRenderData.rdComputeCommandBuffer, VK_PIPELINE_BIND_POINT_COMPUTE,
@@ -4819,8 +4819,8 @@ bool VkRenderer::draw(float deltaTime) {
   vkCmdSetViewport(mRenderData.rdCommandBuffer, 0, 1, &viewport);
   vkCmdSetScissor(mRenderData.rdCommandBuffer, 0, 1, &scissor);
 
-  int worldPosOffset = 0;
-  int skinMatOffset = 0;
+  uint32_t worldPosOffset = 0;
+  uint32_t skinMatOffset = 0;
   for (const auto& model : mModelInstCamData.micModelList) {
     size_t numberOfInstances = mModelInstCamData.micAssimpInstancesPerModel[model->getModelFileName()].size();
     if (numberOfInstances > 0 && model->getTriangleCount() > 0) {
